requirements.txt
Pillow
PyMuPDF
screeninfo
